requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.8
defusedxml==0.7.1
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.55.0
fpdf2==2.8.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-core==0.3.19
langchain-google-genai==2.0.5
langchain-text-splitters==0.3.2
langcodes==3.5.0
langsmith==0.1.143
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
numpy==1.26.4
openai==1.54.5
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.2
srsly==2.4.8
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
yarl==1.17.2
